Create HowLongToWinBingo.cpp
Nothing too interesting, just a quick simulation I wrote to test a game brought up in one of my classes.

In short, the students can choose 6 of 12 words randomly and put them on a bingo sheet. After the students are ready, I say 'x' words and the players check their sheets as I go. First to bingo wins, as the game normally works.